move most of room avatar handling to Avatar, to reuse in editor pills
